Core: Move disable-selection into its own module
